+++ app/src/main/java/com/nktnet/webview_kiosk/ui/components/common/settings/EditableSettingItem.kt
@@ -1,15 +1,12 @@
-package com.nktnet.webview_kiosk.ui.components.common
+package com.nktnet.webview_kiosk.ui.components.common.settings
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.*
 import androidx.compose.material3.DividerDefaults
-import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Edit
-import androidx.compose.material.icons.filled.Info
+import androidx.compose.material3.HorizontalDivider
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -72,9 +69,9 @@ fun EditableSettingItem(
             )
         }
         HorizontalDivider(
+            modifier = Modifier.padding(top = 8.dp),
             thickness = DividerDefaults.Thickness,
-            color = DividerDefaults.color,
-            modifier = Modifier.padding(top = 8.dp)
+            color = DividerDefaults.color
         )
     }
 
@@ -100,67 +97,4 @@ fun EditableSettingItem(
             }
         )
     }
-}
-
-@Composable
-fun EditFieldDialog(
-    title: String,
-    infoText: String,
-    placeholder: String,
-    value: String,
-    onValueChange: (String) -> Unit,
-    isMultiline: Boolean = false,
-    isError: Boolean = false,
-    onDismiss: () -> Unit,
-    onSave: () -> Unit
-) {
-    var showInfoDialog by remember { mutableStateOf(false) }
-
-    if (showInfoDialog) {
-        AlertDialog(
-            onDismissRequest = { showInfoDialog = false },
-            title = { Text("Info") },
-            text = {
-                Column(modifier = Modifier.verticalScroll(rememberScrollState())) {
-                    Text(infoText.replace("\t", "    "))
-                }
-            },
-            confirmButton = {
-                TextButton(onClick = { showInfoDialog = false }) { Text("OK") }
-            }
-        )
-    }
-
-    AlertDialog(
-        onDismissRequest = onDismiss,
-        title = {
-            Row(verticalAlignment = Alignment.CenterVertically) {
-                Text(title, modifier = Modifier.weight(1f))
-                Icon(
-                    imageVector = Icons.Default.Info,
-                    contentDescription = "Info",
-                    modifier = Modifier
-                        .size(20.dp)
-                        .clickable { showInfoDialog = true }
-                        .align(Alignment.CenterVertically)
-                )
-            }
-        },
-        text = {
-            OutlinedTextField(
-                value = value,
-                onValueChange = onValueChange,
-                isError = isError,
-                placeholder = { Text(placeholder) },
-                singleLine = !isMultiline,
-                modifier = if (isMultiline) Modifier.height(200.dp) else Modifier.fillMaxWidth()
-            )
-        },
-        confirmButton = {
-            TextButton(onClick = onSave) { Text("Save") }
-        },
-        dismissButton = {
-            TextButton(onClick = onDismiss) { Text("Cancel") }
-        }
-    )
 }
